Handle Price

+++ src/main/java/com/AlTaraf/Booking/Entity/unit/availableArea/RoomDetailsForAvailableArea.java
@@ -29,10 +29,10 @@ public class RoomDetailsForAvailableArea {
     private int roomNumber;
 
    @Column(name = "NEW_PRICE")
-   private BigDecimal newPrice;
+   private int newPrice;
 
    @Column(name = "OLD_PRICE")
-   private BigDecimal oldPrice;
+   private int oldPrice;
 
     @Column(name = "ADULTS_ALLOWED")
     private int adultsAllowed;
